pyproject.toml
[tool.poetry]
name = "buckystyle_league"
version = "0.1.0"
description = "Buckystyle League Championship Series Data Analysis and Discord Bot"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"

pandas = "^1.5.3"
numpy = "^1.24.2"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^1.4.47"
psycopg2-binary = "^2.9.6"
matplotlib = "^3.7.1"
plotly = "^5.14.1"
dataframe-image = "^0.5.1"

discord.py = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
